update index.js
This allows the same index.js file to be used in Alteryx and for testing in the browser. It uses the presence of a window.Alteryx.browser variable to detect if the script is running inside Alteryx CEF or not. This can be cleaned up considerably, and I will do that during a refactor.